+++ lib/system/freertos/CMakeLists.txt
@@ -21,3 +21,10 @@ if (EXISTS ${CMAKE_CURRENT_SOURCE_DIR}/${PROJECT_MACHINE})
   add_subdirectory(${PROJECT_MACHINE})
 endif (EXISTS ${CMAKE_CURRENT_SOURCE_DIR}/${PROJECT_MACHINE})
 
+if ("${PROJECT_MACHINE}" STREQUAL "zynqmp_a53" OR
+    "${PROJECT_MACHINE}" STREQUAL "zynqmp_a72" OR
+    "${PROJECT_MACHINE}" STREQUAL "zynqmp_a78" OR
+    "${PROJECT_MACHINE}" STREQUAL "zynqmp_r5" OR
+    "${PROJECT_MACHINE}" STREQUAL "zynq7")
+add_subdirectory(xlnx)
+endif()
