+++ lib/system/freertos/sys.h
@@ -19,7 +19,11 @@
 
 #include <metal/errno.h>
 
+#ifdef XLNX_PLATFORM
+#include <metal/system/freertos/xlnx/sys.h>
+#else
 #include "./@PROJECT_MACHINE@/sys.h"
+#endif
 
 #ifdef __cplusplus
 extern "C" {
