+++ lib/system/freertos/xlnx/CMakeLists.txt
@@ -1,4 +1,5 @@
 collect (PROJECT_LIB_HEADERS sys.h)
 
 collect (PROJECT_LIB_SOURCES irq.c)
+collect (PROJECT_LIB_SOURCES sys.c)
